+++ app/src/main/java/eu/trails2education/trails/ContentActivity.java
@@ -16,13 +16,12 @@ import org.json.JSONObject;
 import eu.trails2education.trails.database.Content;
 import eu.trails2education.trails.database.InterestPoint;
 import eu.trails2education.trails.json.InterestPointJSON;
-import eu.trails2education.trails.path.InterestPointUtils;
+import eu.trails2education.trails.network.InterestPointUtils;
 import eu.trails2education.trails.views.ContentSelectionAdapter;
 
 public class ContentActivity extends AppCompatActivity {
 
     RecyclerView contentList;
-    ContentSelectionAdapter adapter;
 
     private InterestPoint interestPoint;
 
@@ -40,19 +39,13 @@ public class ContentActivity extends AppCompatActivity {
             }
         });
 
-        // Fill the title and description
-        //interestPoint = (InterestPoint) getIntent().getSerializableExtra("InterestPoint");
-
         final int interestPointID = (int)getIntent().getExtras().getLong("InterestPointID");
-        // Read the path from the network
         InterestPointUtils.readInterestPointFromNetwork(this, new Response.Listener<JSONObject>() {
             @Override
             public void onResponse(JSONObject response) {
-                Log.e("RESPONSE" , response.toString());
                 try {
                     interestPoint = InterestPointJSON.createInterestPointFromJSON(ContentActivity.this, response.getJSONArray("posts").getJSONObject(0), 0);
                     fillViews();
-                    Log.e("INTEREST POINT LOADED ", "SUCCESS at id: " + interestPointID);
 
                 }catch(Exception e){
                     Log.e("INTEREST POINT LOADING ", "ERROR at id: " + interestPointID);
@@ -63,8 +56,10 @@ public class ContentActivity extends AppCompatActivity {
 
     }
 
+    /**
+     * Updates the views. Network or DB updates may change data in interestPoint.
+     * */
     public void fillViews(){
-        Log.e("CONTENTS", interestPoint.getContents().size() + "");
         if(interestPoint.getContents().size() > 0){
             contentList = (RecyclerView)findViewById(R.id.recyclerView);
             contentList.setAdapter(new ContentSelectionAdapter(interestPoint.getContents())); // Pass the ids for the icons
@@ -72,12 +67,10 @@ public class ContentActivity extends AppCompatActivity {
 
             // Get first subject to populate views
             Content first = interestPoint.getContents().get(0);
-            //Log.e("Loading", first.description);
 
             ((TextView)findViewById(R.id.subject_title)).setText(first.gettitEN());
             ((TextView)findViewById(R.id.subject_content)).setText(first.getdesEN());
         }else{
-            // There is no content!!!!
         }
     }
